move account components into a subdirectory
