+++ src/components/Account/ForgotPassword.jsx
@@ -1,5 +1,5 @@
 import { useState } from 'react'
-import supabase from "../Config/supabaseConfig"
+import supabase from "../../Config/supabaseConfig"
 
 
 export default function ForgotPassword() {
