+++ src/components/Account/LoginSupabase.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { AuthContext } from "./AuthContext";
-import supabase from '../Config/supabaseConfig';
+import { AuthContext } from "../AuthContext";
+import supabase from '../../Config/supabaseConfig';
 
 
 export default function LoginSupabase() {
